Move to Pharos::CommandOptions, add YesOption
